requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.1.4
django-appconf==1.0.6
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-fsm==3.0.0
django-fsm-log==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-nested-routers==0.94.1
filelock==3.16.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
celery[redis]
redis
django-lifecycle==1.0.0
black==23.1.0